% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl.R
\name{vf_read_fl}
\alias{vf_read_fl}
\title{Read FL voter file}
\usage{
vf_read_fl(path_detail, path_history, widen_history = TRUE)
}
\arguments{
\item{path_detail}{path to FL detail file or directory}

\item{path_history}{path to FL history file or directory}
}
\value{
tibble
}
\description{
Read FL voter file
}
\examples{
# TODO
}
