% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{vf_read_nc}
\alias{vf_read_nc}
\title{Read NC voter file}
\usage{
vf_read_nc(path_registration, path_history, ...)
}
\arguments{
\item{path_registration}{path to NC registration file or directory}

\item{path_history}{path to NC history file or directory}

\item{...}{additional arguments passed to \code{readr::read_csv()}}
}
\value{
tibble
}
\description{
Read NC voter file
}
\examples{
# TODO
}
