% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ny.R
\name{vf_read_ny}
\alias{vf_read_ny}
\title{Read NY voter file}
\usage{
vf_read_ny(path, ...)
}
\arguments{
\item{path}{path to NY voter file}

\item{...}{additional arguments passed to \code{readr::read_csv()}}
}
\value{
tibble
}
\description{
Read NY voter file
}
\examples{
# TODO
}
