% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh.R
\name{vf_read_oh}
\alias{vf_read_oh}
\title{Read OH voter file}
\usage{
vf_read_oh(path_registration, ...)
}
\arguments{
\item{path_registration}{path to OH registration file or directory}

\item{...}{additional arguments passed to \code{readr::read_csv()}}
}
\value{
tibble
}
\description{
Read OH voter file
}
\examples{
# TODO
}
