% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ok.R
\name{vf_read_ok}
\alias{vf_read_ok}
\title{Read OK voter file}
\usage{
vf_read_ok(path_registration, ...)
}
\arguments{
\item{path_registration}{path to OK registration file or directory}

\item{...}{additional arguments passed to \code{readr::read_csv()}}
}
\value{
tibble
}
\description{
Read OK voter file
}
\examples{
# TODO
}
